scrape_tools/requirements.txt
scrapy~=2.11.2
spacy~=3.7.6
beautifulsoup4~=4.10.0
gitpython~=3.1.24
pymongo~=4.9.1
